tool: merge oc and kubectl
