+++ pkg/tool/kubectl.go
@@ -13,6 +13,19 @@ import (
 	"k8s.io/kubectl/pkg/scheme"
 )
 
+// Based on https://access.redhat.com/solutions/4870701
+var kubeOpenShiftVersionMap map[string]string = map[string]string{
+	"1.22": "4.9",
+	"1.21": "4.8",
+	"1.20": "4.7",
+	"1.19": "4.6",
+	"1.18": "4.5",
+	"1.17": "4.4",
+	"1.16": "4.3",
+	"1.14": "4.2",
+	"1.13": "4.1",
+}
+
 type Kubectl struct {
 	clientset kubernetes.Interface
 }
@@ -68,4 +81,25 @@ func (k Kubectl) DeleteNamespace(namespace string) error {
 		return err
 	}
 	return nil
+}
+
+func (k Kubectl) GetOcVersion() (string, error) {
+	version, err := k.clientset.Discovery().ServerVersion()
+	if err != nil {
+		return "", err
+	}
+
+	// Relying on Kubernetes version can be replaced after fixing this issue:
+	// https://bugzilla.redhat.com/show_bug.cgi?id=1850656
+	kubeVersion := fmt.Sprintf("%s.%s", version.Major, version.Minor)
+	osVersion, ok := kubeOpenShiftVersionMap[kubeVersion]
+	if !ok {
+		return "", fmt.Errorf("internal error: %q not found in Kubernetes-OpenShift version map", kubeVersion)
+	}
+
+	return osVersion, nil
+}
+
+func GetKubeOpenShiftVersionMap() map[string]string {
+	return kubeOpenShiftVersionMap
 }
